+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 - New view property `hovered-subject` which takes in a `Item` or a `NPC` and passes it to Lookup Anything.
 
 ### Fixed
+- [issue#87](https://github.com/focustense/StardewUI/issues/87)
 - [issue#101](https://github.com/focustense/StardewUI/issues/101)
 
 ## [0.6.2-unofficial-mushymato.2](https://github.com/Mushymato/StardewUI/releases/tag/0.6.2-unofficial-mushymato.2) - 2026-03-18