+++ CHANGELOG.md
@@ -7,9 +7,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 ## [0.6.3-unofficial-mushymato.0](https://github.com/Mushymato/StardewUI/releases/tag/0.6.3-unofficial-mushymato.0) - WIP
 
 ### Added
-- New `scrollable` properties `progress` and `scroll-step`.
-- New view property `hovered-subject` which takes in a `Item` or a `NPC` and passes it to Lookup Anything. [issue#81](https://github.com/focustense/StardewUI/issues/81)
-- New view property `screen-read` which takes a string or a `IScreenReadable` and passes it to Stardew Access.
+- New `scrollable` attributes `progress` for controlling scroll progress, and `scroll-step` for how much to scroll each turn of mouse wheel.
+- New view attribute `hovered-subject` which takes in a `Item` or a `NPC` and passes it to Lookup Anything. [issue#81](https://github.com/focustense/StardewUI/issues/81)
+- New view attribute `screen-read` which takes a string or a `IScreenReadable` and passes it to Stardew Access.
 
 ### Fixed
 - [issue#87](https://github.com/focustense/StardewUI/issues/87)